+++ tests/lib/rules/jsx-quotes.js
@@ -12,6 +12,10 @@
 var rule = require('../../../lib/rules/jsx-quotes');
 var RuleTester = require('eslint').RuleTester;
 
+var DEPRECATION_WARNING = 'The react/jsx-quotes rule is deprecated. Please use the jsx-quotes rule instead.';
+var SINGLEQUOTE_WARNING = 'JSX attributes must use singlequote.';
+var DOUBLEQUOTE_WARNING = 'JSX attributes must use doublequote.';
+
 // -----------------------------------------------------------------------------
 // Tests
 // -----------------------------------------------------------------------------
@@ -19,16 +23,30 @@ var RuleTester = require('eslint').RuleTester;
 var ruleTester = new RuleTester();
 ruleTester.run('jsx-quotes', rule, {
   valid: [
-    {code: '<App foo=\'bar\' />;', options: ['single'], ecmaFeatures: {jsx: true}},
-    {code: '<App foo="bar" />;', options: ['double'], ecmaFeatures: {jsx: true}},
-    {code: '<App foo="ba\'r" />;', options: ['single', 'avoid-escape'], ecmaFeatures: {jsx: true}},
-    {code: '<App foo=\'ba"r\' />;', options: ['double', 'avoid-escape'], ecmaFeatures: {jsx: true}},
-    {code: '<App>foo</App>;', options: ['single'], ecmaFeatures: {jsx: true}}
+    // None, should always trigger at least the deprecation warning
   ],
   invalid: [
-    {code: '<App foo="bar" />;',
-     errors: [{message: 'JSX attributes must use singlequote.'}], options: ['single'], ecmaFeatures: {jsx: true}},
+    {code: '<App />;',
+     errors: [{message: DEPRECATION_WARNING}], ecmaFeatures: {jsx: true}},
     {code: '<App foo=\'bar\' />;',
-     errors: [{message: 'JSX attributes must use doublequote.'}], options: ['double'], ecmaFeatures: {jsx: true}}
+     errors: [{message: DEPRECATION_WARNING}], options: ['single'], ecmaFeatures: {jsx: true}},
+    {code: '<App foo="bar" />;',
+     errors: [{message: DEPRECATION_WARNING}], options: ['double'], ecmaFeatures: {jsx: true}},
+    {code: '<App foo="ba\'r" />;',
+     errors: [{message: DEPRECATION_WARNING}], options: ['single', 'avoid-escape'], ecmaFeatures: {jsx: true}},
+    {code: '<App foo=\'ba"r\' />;',
+     errors: [{message: DEPRECATION_WARNING}], options: ['double', 'avoid-escape'], ecmaFeatures: {jsx: true}},
+    {code: '<App>foo</App>;',
+     errors: [{message: DEPRECATION_WARNING}], options: ['single'], ecmaFeatures: {jsx: true}},
+    {code: '<App foo="bar" />;',
+     errors: [
+      {message: DEPRECATION_WARNING},
+      {message: SINGLEQUOTE_WARNING}
+     ], options: ['single'], ecmaFeatures: {jsx: true}},
+    {code: '<App foo=\'bar\' />;',
+     errors: [
+      {message: DEPRECATION_WARNING},
+      {message: DOUBLEQUOTE_WARNING}
+     ], options: ['double'], ecmaFeatures: {jsx: true}}
   ]
 });