+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
+## [Unreleased]
+### Changed
+* Deprecate `jsx-quotes` rule ([#217][])
+
+[Unreleased]: https://github.com/yannickcr/eslint-plugin-react/compare/v3.3.2...HEAD
+[#217]: https://github.com/yannickcr/eslint-plugin-react/issues/217
+
 ## [3.3.2] - 2015-09-10
 ### Changed
 * Add `state` in lifecycle methods for `sort-comp` rule ([#197][] @mathieudutour)